Remove interceptors from configauth interface
Fixes #4582
