+++ config/configauth/serverauth.go
@@ -16,18 +16,8 @@ package configauth // import "go.opentelemetry.io/collector/config/configauth"
 
 import (
 	"context"
-	"errors"
-	"net/http"
-
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/metadata"
 
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/internal/middleware"
-)
-
-var (
-	errMetadataNotFound = errors.New("no request metadata found")
 )
 
 // ServerAuthenticator is an Extension that can be used as an authenticator for the configauth.Authentication option.
@@ -47,90 +37,8 @@ type ServerAuthenticator interface {
 	// on tenancy as determined by the group membership, or passing through the authentication data to the next collector/backend.
 	// The context keys to be used are not defined yet.
 	Authenticate(ctx context.Context, headers map[string][]string) (context.Context, error)
-
-	// GRPCUnaryServerInterceptor is a helper method to provide a gRPC-compatible UnaryServerInterceptor, typically calling the authenticator's Authenticate method.
-	// While the context is the typical source of authentication data, the interceptor is free to determine where the auth data should come from. For instance, some
-	// receivers might implement an interceptor that looks into the payload instead.
-	// Once the authentication succeeds, the interceptor is expected to call the handler.
-	// See https://pkg.go.dev/google.golang.org/grpc#UnaryServerInterceptor.
-	GRPCUnaryServerInterceptor(ctx context.Context, req interface{}, srvInfo *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (interface{}, error)
-
-	// GRPCStreamServerInterceptor is a helper method to provide a gRPC-compatible StreamServerInterceptor, typically calling the authenticator's Authenticate method.
-	// While the context is the typical source of authentication data, the interceptor is free to determine where the auth data should come from. For instance, some
-	// receivers might implement an interceptor that looks into the payload instead.
-	// Once the authentication succeeds, the interceptor is expected to call the handler.
-	// See https://pkg.go.dev/google.golang.org/grpc#StreamServerInterceptor.
-	GRPCStreamServerInterceptor(srv interface{}, stream grpc.ServerStream, srvInfo *grpc.StreamServerInfo, handler grpc.StreamHandler) error
-
-	// HTTPInterceptor is a helper method to provide an HTTP handler responsible for intercepting the incoming HTTP requests, using the
-	// request's meta data as source of data for the authentication. Once the authentication succeeds, the interceptor is expected to call
-	// the next handler.
-	HTTPInterceptor(next http.Handler) http.Handler
 }
 
 // AuthenticateFunc defines the signature for the function responsible for performing the authentication based on the given headers map.
 // See ServerAuthenticator.Authenticate.
 type AuthenticateFunc func(ctx context.Context, headers map[string][]string) (context.Context, error)
-
-// GRPCUnaryInterceptorFunc defines the signature for the function intercepting unary gRPC calls, useful for authenticators to use as
-// types for internal structs, making it easier to mock them in tests.
-// See ServerAuthenticator.GRPCUnaryServerInterceptor.
-type GRPCUnaryInterceptorFunc func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler, authenticate AuthenticateFunc) (interface{}, error)
-
-// GRPCStreamInterceptorFunc defines the signature for the function intercepting streaming gRPC calls, useful for authenticators to use as
-// types for internal structs, making it easier to mock them in tests.
-// See ServerAuthenticator.GRPCStreamServerInterceptor.
-type GRPCStreamInterceptorFunc func(srv interface{}, stream grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler, authenticate AuthenticateFunc) error
-
-// HTTPInterceptorFunc defines the signature for the function intercepting HTTP calls, useful for authenticators to use as
-// types for internal structs, making it easier to mock them in tests.
-type HTTPInterceptorFunc func(handler http.Handler, authenticate AuthenticateFunc) http.Handler
-
-// DefaultGRPCUnaryServerInterceptor provides a default implementation of GRPCUnaryInterceptorFunc, useful for most authenticators.
-// It extracts the headers from the incoming request, under the assumption that the credentials will be part of the resulting map.
-func DefaultGRPCUnaryServerInterceptor(ctx context.Context, req interface{}, _ *grpc.UnaryServerInfo, handler grpc.UnaryHandler, authenticate AuthenticateFunc) (interface{}, error) {
-	headers, ok := metadata.FromIncomingContext(ctx)
-	if !ok {
-		return nil, errMetadataNotFound
-	}
-
-	ctx, err := authenticate(ctx, headers)
-	if err != nil {
-		return nil, err
-	}
-
-	return handler(ctx, req)
-}
-
-// DefaultGRPCStreamServerInterceptor provides a default implementation of GRPCStreamInterceptorFunc, useful for most authenticators.
-// It extracts the headers from the incoming request, under the assumption that the credentials will be part of the resulting map.
-func DefaultGRPCStreamServerInterceptor(srv interface{}, stream grpc.ServerStream, _ *grpc.StreamServerInfo, handler grpc.StreamHandler, authenticate AuthenticateFunc) error {
-	ctx := stream.Context()
-	headers, ok := metadata.FromIncomingContext(ctx)
-	if !ok {
-		return errMetadataNotFound
-	}
-
-	ctx, err := authenticate(ctx, headers)
-	if err != nil {
-		return err
-	}
-
-	wrapped := middleware.WrapServerStream(stream)
-	wrapped.WrappedContext = ctx
-	return handler(srv, wrapped)
-}
-
-// DefaultHTTPInterceptor provides a default implementation of HTTPInterceptorFunc, useful for most authenticators.
-// It passes the headers from the incoming request as it is, under the assumption that the credentials are part of it.
-func DefaultHTTPInterceptor(next http.Handler, authenticate AuthenticateFunc) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		ctx, err := authenticate(r.Context(), r.Header)
-		if err != nil {
-			http.Error(w, http.StatusText(http.StatusUnauthorized), http.StatusUnauthorized)
-			return
-		}
-
-		next.ServeHTTP(w, r.WithContext(ctx))
-	})
-}
