+++ config/configgrpc/configgrpc.go
@@ -17,6 +17,7 @@ package configgrpc // import "go.opentelemetry.io/collector/config/configgrpc"
 import (
 	"context"
 	"crypto/tls"
+	"errors"
 	"fmt"
 	"net"
 	"strings"
@@ -31,6 +32,7 @@ import (
 	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/encoding/gzip"
 	"google.golang.org/grpc/keepalive"
+	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/peer"
 
 	"go.opentelemetry.io/collector/client"
@@ -56,6 +58,8 @@ var (
 		CompressionSnappy: snappy.Name,
 		CompressionZstd:   zstd.Name,
 	}
+
+	errMetadataNotFound = errors.New("no request metadata found")
 )
 
 // Allowed balancer names to be set in grpclb_policy to discover the servers.
@@ -337,8 +341,12 @@ func (gss *GRPCServerSettings) ToServerOption(host component.Host, settings comp
 			return nil, err
 		}
 
-		uInterceptors = append(uInterceptors, authenticator.GRPCUnaryServerInterceptor)
-		sInterceptors = append(sInterceptors, authenticator.GRPCStreamServerInterceptor)
+		uInterceptors = append(uInterceptors, func(ctx context.Context, req interface{}, info *grpc.UnaryServerInfo, handler grpc.UnaryHandler) (resp interface{}, err error) {
+			return authUnaryServerInterceptor(ctx, req, info, handler, authenticator.Authenticate)
+		})
+		sInterceptors = append(sInterceptors, func(srv interface{}, ss grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
+			return authStreamServerInterceptor(srv, ss, info, handler, authenticator.Authenticate)
+		})
 	}
 
 	// Enable OpenTelemetry observability plugin.
@@ -394,4 +402,35 @@ func contextWithClient(ctx context.Context) context.Context {
 		cl.Addr = p.Addr
 	}
 	return client.NewContext(ctx, cl)
+}
+
+func authUnaryServerInterceptor(ctx context.Context, req interface{}, _ *grpc.UnaryServerInfo, handler grpc.UnaryHandler, authenticate configauth.AuthenticateFunc) (interface{}, error) {
+	headers, ok := metadata.FromIncomingContext(ctx)
+	if !ok {
+		return nil, errMetadataNotFound
+	}
+
+	ctx, err := authenticate(ctx, headers)
+	if err != nil {
+		return nil, err
+	}
+
+	return handler(ctx, req)
+}
+
+func authStreamServerInterceptor(srv interface{}, stream grpc.ServerStream, _ *grpc.StreamServerInfo, handler grpc.StreamHandler, authenticate configauth.AuthenticateFunc) error {
+	ctx := stream.Context()
+	headers, ok := metadata.FromIncomingContext(ctx)
+	if !ok {
+		return errMetadataNotFound
+	}
+
+	ctx, err := authenticate(ctx, headers)
+	if err != nil {
+		return err
+	}
+
+	wrapped := middleware.WrapServerStream(stream)
+	wrapped.WrappedContext = ctx
+	return handler(srv, wrapped)
 }
